+++ examples/portal.rs
@@ -88,6 +88,7 @@ fn setup(mut commands: Commands, mut effects: ResMut<Assets<EffectAsset>>) {
             })
             .render(SizeOverLifetimeModifier {
                 gradient: size_gradient1,
+                screen_space_size: false,
             })
             .render(OrientAlongVelocityModifier),
     );